refactor: split tag matrix coverage logic
